+++ src/carli/experimental_output.h
@@ -3,6 +3,7 @@
 
 #include <cstddef>
 #include <functional>
+#include <inttypes.h>
 
 #include "linkage.h"
 
@@ -10,9 +11,9 @@ namespace Carli {
 
   class CARLI_LINKAGE Experimental_Output {
   public:
-    Experimental_Output(const size_t &print_every = 1);
+    Experimental_Output(const int64_t &print_every = 1);
 
-    void print(const size_t &total_steps, const size_t &episode_number, const size_t &step_count, const double &reward, const bool &done, const size_t &q_value_count);
+    void print(const int64_t &total_steps, const int64_t &episode_number, const int64_t &step_count, const double &reward, const bool &done, const int64_t &q_value_count);
 
   private:
     void reset_stats();
@@ -20,9 +21,9 @@ namespace Carli {
     double m_cumulative_reward = 0.0;
     double m_simple_reward = 0.0;
 
-    size_t m_print_every;
+    int64_t m_print_every;
 
-    size_t m_print_count;
+    int64_t m_print_count;
 
     double m_cumulative_min;
     double m_cumulative_mean;
